LPD-55202 Move to the common services folder
