+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/VocabularyFDSPropsTransformer.ts
@@ -5,7 +5,7 @@
 
 import {IInternalRenderer} from '@liferay/frontend-data-set-web';
 
-import {IVocabulary} from '../categorization/types/IVocabulary';
+import {IVocabulary} from '../../types/IVocabulary';
 import {openGenericFDSDeleteConfirmationModal} from '../util/GenericOpenModalUtil';
 import VocabularyRenderer from './cell_renderers/VocabularyRenderer';
